requirements.txt
Django==3.0.1
django-choices==1.7.1
uvicorn==0.10.8
djangorestframework==3.11.0
